netfilter: nf_ct_helper: allocate 16 bytes for the helper and policy names
This patch modifies the struct nf_conntrack_helper to allocate
the room for the helper name. The maximum length is 16 bytes
(this was already introduced in 2.6.24).

For the maximum length for expectation policy names, I have
also selected 16 bytes.

This patch is required by the follow-up patch to support
user-space connection tracking helpers.

Change-Id: I529634d3cfa53abab6dd7c2b9d3517dedbdcf69f
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>