msm: netfilter: NATTYPE Refresh Timer Changes.
When IPA is present, all the data packets go through IPA
and as a result NATTYPE entry timeout will not be refreshed
and eventually it times out. IPA periodically refreshes
the connection tracking entry timeout. So make changes to refresh
the NATTYPE entry timeout from the connection tracking module.

Change-Id: I5861427990af4bfd6046d21809a778409d0d8d5f
Signed-off-by: Niranjan Reddy Dumbala <ndumba@codeaurora.org>

Conflicts:
	include/linux/netfilter_ipv4/ipt_NATTYPE.h
	net/ipv4/netfilter/ipt_NATTYPE.c
	net/netfilter/nf_conntrack_core.c