cpufreq: conservative: add a two step frequency scale up
Add a per_cpu twostep_counter that tracks if cpuN is above 70% load. If this
happens for 2 samples in a row and the 3rd sample the cpuN is above
up_threshold then we allow the governor to scale up to policy->max.
Before the counter reaches a value of 2, if the load is above 70% but below
up_threshold we scale up to max / 2.

We reset the counter if it scales to max, of if it scales down
regardless of the frequency (it will be always below max / 2 anyway)

Benchmarks did not show any degradation on its numbers, yet we see lower
usage of policy->max and higher time spent on max / 2 which will decrease
power consumption during regular usage.

Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>

Conflicts:
	drivers/cpufreq/cpufreq_conservative.c

Change-Id: I27d841930a0590ce6c9ca2acc7c3dca1442af157